LPD-31503 Moves MasterPagesPage.ts since we are going to reuse it
