+++ modules/test/playwright/tests/layout-page-template-admin-web/masterPagesAdmin.spec.ts
@@ -7,10 +7,10 @@ import {Locator, expect, mergeTests} from '@playwright/test';
 
 import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
+import {masterPagesPagesTest} from '../../fixtures/masterPagesPagesTest';
 import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
 import {pagesAdminPagesTest} from '../../fixtures/pagesAdminPagesTest';
 import getRandomString from '../../utils/getRandomString';
-import {masterPagesPagesTest} from './fixtures/masterPagesPagesTest';
 
 export const test = mergeTests(
 	pagesAdminPagesTest,
